Rename parseutils pkg to strvals
