--- lib/parseutils/doc.go
@@ -1,2 +1,0 @@
-// Package parseutils provides text parsing utilities.
-package parseutils
