+++ lib/strvals/doc.go
@@ -1,0 +1,3 @@
+// Package strvals provides parsing utilities for strval lines, which follow
+// the format: name=value,topname.subname=value.
+package strvals
