+++ lib/strvals/tokenizer.go
@@ -1,4 +1,4 @@
-package parseutils
+package strvals
 
 import "fmt"
